% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Boyle}
\alias{Boyle}
\title{Boyle data set}
\format{
A data frame with 25 rows and 2 variables:
\describe{
  \item{volume}{the number of equal spaces in the shorter leg, that contained the same parcel of air diversely extended}
  \item{pressure}{the pressure sustained by the included air}
}
}
\source{
\url{https://www.chemteam.info/GasLaw/Gas-Boyle-Data.html}
}
\usage{
data(Boyle)
}
\description{
Robert Boyle's data set
}
\keyword{datasets}
