% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Death}
\alias{Death}
\title{Death rate and related variables for Korean districts}
\format{
A data frame with 258 rows and 9 variables:
\describe{
  \item{region}{region ID}
  \item{year}{year}
  \item{regpop}{registered population (end of year)}
  \item{death}{number of registered deaths}
  \item{drink}{percentage of drinkers (more than once in a month)}
  \item{smoke}{percentage of smokers (smoker = has smoked 100+ cigarettes and currently smoking)}
  \item{aged}{percentage of those aged 65 and over}
  \item{vehipc}{number of vehicles per person}
  \item{deathrate}{= death/regpop*1000}
}
}
\source{
Statistics Korea
}
\usage{
data(Death)
}
\description{
Death rate and related variables for Korean districts for 2008-2010
}
\keyword{datasets}
