% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ekc}
\alias{Ekc}
\title{CO2 emissions}
\format{
A data frame with 183 rows and 4 variables:
\describe{
  \item{ccode}{country code}
  \item{cname}{country name}
  \item{gdppcppp}{GDP per capital, ppp adjusted (USD)}
  \item{co2pc}{CO2 emissions per capita (ton)}
}
}
\source{
\url{http://wdi.worldbank.org}
}
\usage{
data(Ekc)
}
\description{
CO2 emissions per capita and GDP per capita in 2005
}
\author{
\packageAuthor{loedata}
}
\keyword{datasets}
