% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hcons}
\alias{Hcons}
\title{Household consumption shares}
\format{
A data frame with 6723 rows of 3 variables:
\describe{
  \item{age}{age of household head}
  \item{comm}{share of consumption for communication in \%}
  \item{rec}{share of consumption for recreation in \%}
}
}
\source{
Korea Household Income and Expenditure Survey 2014 \url{http://kostat.go.kr/portal/eng/surveyOutline/6/1/index.static}
}
\usage{
data(Hcons)
}
\description{
Household consumption shares of communication and recreation sector in Korean Household Income and Expenditure Survey 2014
}
\seealso{
\code{\link{Hies}}
}
\keyword{datasets}
