% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hies}
\alias{Hies}
\title{Household Income and Expenditure Survey 2016}
\format{
A data frame with 1368 rows of 26 variables:
\describe{
  \item{year}{year of survey, =2016}
  \item{famsize}{number of family members}
  \item{empnum}{number of employed members}
  \item{age}{age of household head}
  \item{emp}{1 if head is employed}
  \item{ownhouse}{1 if own house}
  \item{weight}{cross sectional weight}
  \item{inc}{household monthly income}
  \item{haspinc}{1 if has income from properties}
  \item{totexp}{household total monthly expenditure}
  \item{cons}{household monthly consumption}
  \item{cons01}{household monthly consumption in section 01}
  \item{cons02}{household monthly consumption in section 02}
  \item{cons03}{household monthly consumption in section 03}
  \item{cons04}{household monthly consumption in section 04}
  \item{cons05}{household monthly consumption in section 05}
  \item{cons06}{household monthly consumption in section 06}
  \item{cons07}{household monthly consumption in section 07}
  \item{cons08}{household monthly consumption in section 08}
  \item{cons09}{household monthly consumption in section 09}
  \item{cons10}{household monthly consumption in section 10}
  \item{cons11}{household monthly consumption in section 11}
  \item{cons12}{household monthly consumption in section 12}
  \item{propens}{propensity to consume (=cons/inc)}
  \item{educ}{years of head's education}
  \item{female}{1 if head is female}
}
}
\source{
\url{http://kostat.go.kr/portal/eng/surveyOutline/6/1/index.static}
}
\usage{
data(Hies)
}
\description{
A subset (30 <= age <= 39) of Korea Household Income and Expenditure Survey 2016
}
\seealso{
\code{\link{Hcons}}
}
\author{
\packageAuthor{loedata}
}
\keyword{datasets}
