% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hmda}
\alias{Hmda}
\title{The Boston HMDA data set}
\format{
A data frame with 2381 rows of 13 variables:
\describe{
  \item{dir}{debt payments to total income ratio}
  \item{hir}{housing expenses to income ratio}
  \item{lvr}{ratio of size of loan to assessed value of propensity}
  \item{ccs}{consumer credit score from 1 to 6 (a low value being a good score)}
  \item{mcs}{mortgage credit score from 1 to 4 (a low value being a good score)}
  \item{pbcr}{1 if public bad credit score}
  \item{dmi}{1 if denied mortgage insurance}
  \item{self}{1 if self employed}
  \item{single}{1 if the applicant is single}
  \item{uria}{1989 Massachusetts unemployment rate in the applicant's industry}
  \item{condominium}{1 if unit is a condominium}
  \item{black}{1 if the applicant is black}
  \item{deny}{1 if mortgage application denied}
}
}
\source{
Hmda data in the Ecdat package
}
\usage{
data(Hmda)
}
\description{
The Boston HMDA data set in the Ecdat package, with yes/no converted to 1/0
}
\keyword{datasets}
