% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ivdata}
\alias{Ivdata}
\title{Artificial data for studying IV estimation}
\format{
A data frame with 100 rows of 5 variables:
\describe{
  \item{y}{y variable}
  \item{x1}{x1 variable}
  \item{x2}{x2 variable}
  \item{z2a}{z2a variable}
  \item{z2b}{z2b variable}
}
}
\usage{
data(Ivdata)
}
\description{
Artificial data for studying IV estimation
}
\keyword{datasets}
