% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Klosa}
\alias{Klosa}
\title{KLoSA wave 4}
\format{
A data frame with 2153 rows of 45 variables:
\describe{
  \item{pid}{personal ID}
  \item{wave}{= (year-2006)/2 + 1}
  \item{male}{1 if male}
  \item{educ}{years of education}
  \item{age}{age}
  \item{married}{1 if married, 0 otherwise}
  \item{childnum}{number of children}
  \item{hsize}{number of housemates}
  \item{region}{region type, one of \code{"big city"}, \code{"small city"}, and \code{"town"}}
  \item{htype}{type of residential facility, either \code{"dwelling"} or \code{"apartment"}}
  \item{religion}{1 if has religion}
  \item{meeting1}{1 if in religious meeting groups}
  \item{meeting2}{1 if in social gathering groups}
  \item{meeting3}{1 if in leisure/sports groups, etc.}
  \item{meeting4}{1 if in union/fraternity groups, etc.}
  \item{meeting5}{1 if in volunteer service groups}
  \item{meeting6}{1 if in political/civic/interest groups}
  \item{health}{health conditions, one of \code{"excellent"},
   \code{"above average"}, \code{"average"}, \code{"below average"}, and \code{"poor"}}
  \item{hlth}{1=poor, 2=below average, 3=average, 4=above average, 5=excellent}
  \item{hlth3}{1=health above average, 0=average, -1=below average}
  \item{height}{height in cm}
  \item{weight}{weight in kg}
  \item{exercise}{period of regular exercise; 0=do not regularly exercise,
   1=0~3mo, 2=4~6mo, 3=7mo~1yr, 4=1~2yr, 5=3~4yr, 6=5~6yr, 7=7+yr}
  \item{bmi}{BMI}
  \item{smoke}{# of cigarettes smoked per day}
  \item{working}{1 if working}
  \item{jobtype}{job type; one of waged employee, self-employed, unemployed, unpaid family worker}
  \item{jobseeking}{1 if seeking a job}
  \item{receive}{amount received from children last year (10k KRW)}
  \item{give}{amount given to children last year (10k KRW)}
  \item{poketm}{regular pocket money received from children (10k KRW)}
  \item{satisfy1}{satisfaction about health conditions}
  \item{satisfy2}{satisfaction about economic conditions}
  \item{satisfy3}{satisfaction about relationship with spouse}
  \item{satisfy4}{satisfaction about relationship with children}
  \item{satisfy5}{satisfaction in comparison to others in the same age group (out of 100)}
  \item{travel1}{number of travels last year}
  \item{travel2}{expenditure on travel (10k KRW)}
  \item{culture1}{number of cultural activities}
  \item{culture2}{expenditure on cultural activities}
  \item{hobby1}{hours for hobbies, per month}
  \item{hobby2}{expenditure on hobbies (10k KRW)}
  \item{training1}{hours for self development, per month}
  \item{training2}{expenditure on self development (10k KRW)}
  \item{voluntary}{hours of volunteer service}
}
}
\source{
\url{https://survey.keis.or.kr/klosa/klosa01.jsp}
}
\usage{
data(Klosa)
}
\description{
Korea Longitudinal Study of Aging wave 4 (2012)
}
\keyword{datasets}
