% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ksalary}
\alias{Ksalary}
\title{Average salary}
\format{
A data frame with 1636 rows and 10 variables:
\describe{
  \item{seqno}{sequential number}
  \item{market}{"kospi" or "kosdaq"}
  \item{sales}{sales in Bil. KRW}
  \item{profit}{profit in Bil. KRW}
  \item{sector}{sector (character)}
  \item{emp}{number of employees}
  \item{avgsal}{average salary in Mil. KRW}
  \item{avgtenure}{average years of tenure}
  \item{kospi}{=1 if KOSPI}
  \item{kosdaq}{=1 if KOSDAQ}
}
}
\source{
\url{https://blog.naver.com/naamoo01/130185489128}
}
\usage{
data(Ksalary)
}
\description{
Average salary for Korean firms in 2012
}
\author{
\packageAuthor{loedata}
}
\keyword{datasets}
