% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loedata}
\alias{loedata}
\alias{loedata-package}
\title{Data package for Lectures on Econometrics}
\description{
This package contains data sets for Lectures on Econometrics by Chirok Han
}
\seealso{
\code{help(package="loedata")}
}
\author{
\packageAuthor{loedata}
}
