% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pubserv}
\alias{Pubserv}
\title{Public servants and financial independence}
\format{
A data frame with 86 rows of 3 variables:
\describe{
  \item{gun}{name of gun}
  \item{servpc}{number of public servants per 1000 pop}
  \item{finind}{financial independence index, = (local tax + other income)/budget * 100}
}
}
\source{
\url{http://kostat.go.kr/}
}
\usage{
data(Pubserv)
}
\description{
Korean regional public servants and financial independence in 2010
}
\author{
\packageAuthor{loedata}
}
\keyword{datasets}
