% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Regko}
\alias{Regko}
\title{Korean regional data (2014-2016 averages)}
\format{
A data frame with 264 rows of 23 variables:
\describe{
  \item{id}{ID of region}
  \item{metro}{Metropolitan region name (metro cities and provinces)}
  \item{region}{Region name}
  \item{type}{1=si (non-metropolitan cities), 2=gun, 3=gu in metro cities and provinces}
  \item{grdp}{gross regional GDP}
  \item{regpop}{registered population (end of year, by registration record, excl. foreigners)}
  \item{popgrowth}{population growth}
  \item{eq5d}{the EQ-5D health index}
  \item{deaths}{number of registered deaths}
  \item{drink}{\% of drinkers}
  \item{hdrink}{\% of high-risk drinkers}
  \item{smoke}{\% of smokers}
  \item{aged}{\% of aged 65 and over}
  \item{divorce}{# of divorces per 1000 pop}
  \item{medrate}{# of medical beds per 1000 pop}
  \item{gcomp}{gender composition # men / 100 women}
  \item{vehipc}{# of vehicles per person}
  \item{accpv}{# of accidents per 1000 vehicles}
  \item{dumppc}{waste dump per person, kg/day}
  \item{stratio}{# of students per teacher}
  \item{deathrate}{# of deaths per 100,000 mid-year population (not equal to deaths/regpop*1e5)}
  \item{nbirth}{# of births}
  \item{cbrate}{Crude birth rate (per 1,000 pop)}
  \item{tfrate}{Total fertility rate}
  \item{pctmale}{=gcmp/(gcomp+100)*100, \% of male}
  \item{accpc}{=vehipc*accpv, # of accidents per 1000 pop}
}
}
\source{
\url{http://kostat.go.kr/}
}
\usage{
data(Regko)
}
\description{
Korean regional data for 2014-2016 average
}
\seealso{
\code{RegkoPanel}
}
\keyword{datasets}
